refactor: normalize premium-analytics widget CSS modules to style.module.css
