+++ projects/packages/premium-analytics/widgets/emails/style.module.css
