+++ projects/packages/premium-analytics/widgets/hello-world/render.tsx
@@ -1,8 +1,11 @@
 /**
- * WordPress dependencies
+ * External dependencies
  */
 import { __ } from '@wordpress/i18n';
 import { Stack, Text } from '@wordpress/ui';
+/**
+ * Internal dependencies
+ */
 import styles from './style.module.css';
 import type { HelloWorldAttributes } from './widget';
 import type { WidgetRenderProps } from '@wordpress/widget-primitives';
@@ -13,9 +16,9 @@ type HelloWorldRenderProps = WidgetRenderProps< HelloWorldAttributes >;
  * Renders the Hello World widget.
  *
  * @param {HelloWorldRenderProps} props - Component props
- * @return {React.ReactNode} The rendered widget.
+ * @return The rendered widget.
  */
-export default function HelloWorld( { attributes = {} }: HelloWorldRenderProps ): React.ReactNode {
+export default function HelloWorld( { attributes = {} }: HelloWorldRenderProps ) {
 	return (
 		<Stack align="center" justify="center" className={ styles.root }>
 			<Text variant="heading-2xl" render={ <h2 /> }>
